# Make CBlock a vector of shared_ptr of CTransactions