… src/miner.cpp
… src/primitives/transaction.cpp
… src/primitives/transaction.h
… src/rpc/blockchain.cpp
… src/rpc/mining.cpp
… src/rpc/rawtransaction.cpp
… src/serialize.h
… src/test/blockencodings_tests.cpp
… src/test/mempool_tests.cpp
… src/test/merkle_tests.cpp
… src/test/miner_tests.cpp
… src/test/pmt_tests.cpp
… src/test/policyestimator_tests.cpp
… src/test/test_bitcoin.cpp
… src/test/test_bitcoin.h
… src/txmempool.cpp
… src/txmempool.h
… src/wallet/wallet.cpp