@sipa 2016-11-11
I created a branch with some follow-up *shared_ptr all the things* and other optimizations: https://github.com/sipa/bitcoin/commits/sharedblock2