@sipa 2016-11-16
Benchmark results (25 reindex-chainstates, default dbcache, otherwise unloaded 24-core machine, measured by -debug=bench) on both master and #9125+#8580+#8589+a few shared_ptr optimizations I haven't PR'ed yet:
* Average: master 1490s, shared_ptr 1442s

Best master run:
```
2016-11-16 12:27:14   - Load block from disk: 1.25ms [273.57s]
2016-11-16 12:27:14     - Sanity checks: 0.14ms [70.95s]
2016-11-16 12:27:14     - Fork checks: 0.02ms [80.76s]
2016-11-16 12:27:14       - Connect 64 transactions: 11.06ms (0.173ms/tx, 0.020ms/txin) [711.67s]
2016-11-16 12:27:14     - Verify 550 txins: 11.09ms (0.020ms/txin) [718.24s]
2016-11-16 12:27:14     - Index writing: 0.03ms [5.25s]
2016-11-16 12:27:14     - Callbacks: 0.01ms [3.75s]
2016-11-16 12:27:14   - Connect total: 11.34ms [896.52s]
2016-11-16 12:27:14   - Flush: 0.26ms [86.70s]
2016-11-16 12:27:14   - Writing chainstate: 0.02ms [112.45s]
2016-11-16 12:27:14   - Connect postprocess: 0.42ms [86.88s]
2016-11-16 12:27:14 - Connect block: 13.30ms [1456.12s]
```

Best shared_ptr run:
```
2016-11-16 10:17:40   - Load block from disk: 1.61ms [283.93s]
2016-11-16 10:17:40     - Sanity checks: 0.18ms [76.82s]
2016-11-16 10:17:40     - Fork checks: 0.03ms [83.22s]
2016-11-16 10:17:40       - Connect 64 transactions: 17.43ms (0.272ms/tx, 0.032ms/txin) [717.76s]
2016-11-16 10:17:40     - Verify 550 txins: 17.47ms (0.032ms/txin) [724.51s]
2016-11-16 10:17:40     - Index writing: 0.02ms [5.34s]
2016-11-16 10:17:40     - Callbacks: 0.02ms [3.85s]
2016-11-16 10:17:40   - Connect total: 17.81ms [911.36s]
2016-11-16 10:17:40   - Flush: 0.40ms [87.91s]
2016-11-16 10:17:40   - Writing chainstate: 0.03ms [112.31s]
2016-11-16 10:17:40   - Connect postprocess: 0.08ms [25.61s]
2016-11-16 10:17:40 - Connect block: 19.94ms [1421.12s]
```