@sipa 2016-11-17
@jtimon I believe all of them are used in successor PRs.

@theuni That overload won't work for the arguments that take a
CTransactionRef.