@sipa 2016-11-17
@jtimon std::move just casts to an T&& argument. It's to indicate that the
called function may destroy the object.