@theuni 2016-11-17
@sipa I'm unsure why you'd want to do that? Why wouldn't you just
```
CTransactionRef a(MakeTransactionRef());
CTransactionRef b(a);
CTransactionRef c(std::move(b));
```
?