@sipa 2016-11-18
Rebased, and addressed two nits by @Theuni:
* Added the memory usage for the control block of the vtx entries in CBlock.
* Changed deserialize_t to deserialize_type to avoid colliding with the reserved _t suffix.