@sipa 2016-11-19
Some more benchmarks (fastest of 28 reindexes each, 300 MB dbcache, reindex chainstate up to block 295000):
* master: 1462.4s
* this PR: 1413.6s
* #8580: 1412.2s
* #8589: 1420.5s
* https://github.com/sipa/bitcoin/commits/sharedblock2: 1410.3s