@sipa 2016-11-20
Addressed all of @TheBlueMatt's nits.